implementaciones capa logica

+++ src/main/java/edu/eci/cvds/samples/services/ServiciosAlquiler.java
@@ -11,7 +11,7 @@ public interface ServiciosAlquiler {
 
     public abstract int valorMultaRetrasoxDia(int itemId);
 
-    public abstract Cliente consultarCliente(long docu) throws ExcepcionServiciosAlquiler;
+    public abstract Cliente consultarCliente(int docu) throws ExcepcionServiciosAlquiler;
 
     /**
      * @obj Consultar los items que tenga en su poder un cliente
